+++ pybamm/solvers/c_solvers/idaklu/CasadiSolverOpenMP_solvers.cpp
@@ -1,0 +1,1 @@
+#include "CasadiSolverOpenMP_solvers.hpp"